# openApi.yaml
openapi: 3.0.3
info:
  title: Commulab System API
  description: >
    API untuk autentikasi, manajemen pasien, sesi konseling,
    chat AI, transkrip, evaluasi, dan laporan statistik.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /reports/transcripts/{session_id}:
    get:
      tags: [Reports]
      summary: Ambil transkrip session
      description: >
        Mengambil semua transkrip percakapan dalam session tertentu.
        User hanya bisa mengakses transkrip session miliknya sendiri.
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID session konseling
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Transkrip berhasil diambil
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transcript_id:
                      type: integer
                      example: 1
                    session_id:
                      type: integer
                      example: 1
                    sender:
                      type: string
                      enum: [user, ai]
                      example: "user"
                    message:
                      type: string
                      example: "Saya merasa cemas belakangan ini"
                    prosody_data:
                      type: object
                      nullable: true
                      properties:
                        pitch:
                          type: number
                        energy:
                          type: number
                        tempo:
                          type: number
                    timestamp:
                      type: string
                      format: date-time
                      example: "2024-12-27T10:15:30.000Z"
              examples:
                success:
                  summary: Transkrip percakapan
                  value:
                    - transcript_id: 1
                      session_id: 1
                      sender: "user"
                      message: "Halo, saya ingin berbicara tentang masalah saya"
                      prosody_data:
                        pitch: 220.5
                        energy: 0.75
                        tempo: 1.2
                      timestamp: "2024-12-27T10:15:30.000Z"
                    - transcript_id: 2
                      session_id: 1
                      sender: "ai"
                      message: "Halo, saya mendengarkan. Ceritakan apa yang sedang kamu rasakan"
                      prosody_data: null
                      timestamp: "2024-12-27T10:15:35.000Z"
        "404":
          description: Session tidak ditemukan atau tidak punya akses
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sesi tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /reports/evaluation/{session_id}:
    get:
      tags: [Reports]
      summary: Ambil evaluasi session
      description: >
        Mengambil data evaluasi session. Jika belum ada evaluasi,
        akan return data default dengan evaluated: false.
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID session konseling
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Data evaluasi
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: integer
                    example: 1
                  user_id:
                    type: integer
                    example: 1
                  empathy_score:
                    type: integer
                    description: Skor empati 0-100
                    example: 85
                  question_score:
                    type: integer
                    description: Skor pertanyaan 0-100
                    example: 78
                  feedback_text:
                    type: string
                    nullable: true
                    example: "Konselor menunjukkan empati yang baik"
                  evaluated:
                    type: boolean
                    description: Status apakah sudah dievaluasi
                    example: true
                  created_at:
                    type: string
                    format: date-time
              examples:
                evaluated:
                  summary: Session sudah dievaluasi
                  value:
                    session_id: 1
                    user_id: 1
                    empathy_score: 85
                    question_score: 78
                    feedback_text: "Konselor menunjukkan empati yang baik"
                    evaluated: true
                    created_at: "2024-12-27T11:00:00.000Z"
                not_evaluated:
                  summary: Session belum dievaluasi
                  value:
                    session_id: 1
                    user_id: 1
                    empathy_score: 0
                    question_score: 0
                    feedback_text: null
                    evaluated: false
        "404":
          description: Session tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sesi tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    post:
      tags: [Reports]
      summary: Simpan evaluasi session
      description: >
        Menyimpan atau update evaluasi manual untuk session.
        Skor empathy dan question harus antara 0-100.
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID session yang akan dievaluasi
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empathy_score:
                  type: integer
                  minimum: 0
                  maximum: 100
                  description: Skor empati (0-100)
                  example: 85
                question_score:
                  type: integer
                  minimum: 0
                  maximum: 100
                  description: Skor pertanyaan (0-100)
                  example: 78
                feedback_text:
                  type: string
                  nullable: true
                  description: Feedback text (opsional)
                  example: "Konselor menunjukkan empati yang baik"
            examples:
              complete:
                summary: Evaluasi lengkap
                value:
                  empathy_score: 85
                  question_score: 78
                  feedback_text: "Konselor menunjukkan empati yang baik"
              minimal:
                summary: Hanya scores
                value:
                  empathy_score: 80
                  question_score: 75
      responses:
        "200":
          description: Evaluasi berhasil disimpan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Evaluasi berhasil disimpan"
                  data:
                    type: object
                    properties:
                      session_id:
                        type: integer
                      empathy_score:
                        type: integer
                      question_score:
                        type: integer
                      feedback_text:
                        type: string
                        nullable: true
              examples:
                success:
                  summary: Evaluasi tersimpan
                  value:
                    success: true
                    message: "Evaluasi berhasil disimpan"
                    data:
                      session_id: 1
                      empathy_score: 85
                      question_score: 78
                      feedback_text: "Konselor menunjukkan empati yang baik"
        "400":
          description: Validasi gagal
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                invalid_empathy:
                  summary: Empathy score tidak valid
                  value:
                    error: "empathy_score harus antara 0-100"
                invalid_question:
                  summary: Question score tidak valid
                  value:
                    error: "question_score harus antara 0-100"
        "404":
          description: Session tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sesi tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
  /reports/{session_id}/analyze:
    post:
      tags: [Reports]
      summary: Analisis otomatis sesi (AI + prosody)
      description: >
        Melakukan analisis otomatis session menggunakan AI dan data prosody.
        Menghasilkan evaluasi otomatis beserta klasifikasi detail setiap pesan.
        Session harus sudah selesai (status: finished) dan belum pernah dievaluasi.
        Analisis menggunakan Python API eksternal untuk klasifikasi ML.
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID session yang akan dianalisis
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Analisis berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Evaluasi berhasil disimpan"
                  evaluation:
                    type: object
                    description: Data evaluasi yang tersimpan
                    properties:
                      session_id:
                        type: integer
                      empathy_score:
                        type: integer
                      question_score:
                        type: integer
                      feedback_text:
                        type: string
                        nullable: true
                  classification_results:
                    type: array
                    description: Hasil klasifikasi per message
                    items:
                      type: object
                      properties:
                        transcript_id:
                          type: integer
                        message:
                          type: string
                        predicted_class:
                          type: string
                          description: Tipe pertanyaan (open/closed/reflection/etc)
                        confidence:
                          type: number
                          description: Confidence score 0-1
                  detailed_analysis:
                    type: object
                    properties:
                      model_statistics:
                        type: object
                        description: Statistik dari ML model
                        properties:
                          open_questions:
                            type: integer
                          closed_questions:
                            type: integer
                          reflections:
                            type: integer
                      total_counselor_messages:
                        type: integer
                        description: Total pesan dari konselor
                      prosody_analysis:
                        type: object
                        nullable: true
                        description: Analisis prosody user
                        properties:
                          avg_pitch:
                            type: number
                          avg_energy:
                            type: number
                          avg_tempo:
                            type: number
                      python_api_used:
                        type: string
                        description: URL Python API yang digunakan
              examples:
                success:
                  summary: Analisis berhasil
                  value:
                    success: true
                    message: "Evaluasi berhasil disimpan"
                    evaluation:
                      session_id: 1
                      empathy_score: 82
                      question_score: 75
                      feedback_text: "Analisis otomatis: Konselor menggunakan 5 pertanyaan terbuka"
                    classification_results:
                      - transcript_id: 2
                        message: "Bagaimana perasaanmu saat ini?"
                        predicted_class: "open_question"
                        confidence: 0.95
                      - transcript_id: 4
                        message: "Apakah kamu merasa cemas?"
                        predicted_class: "closed_question"
                        confidence: 0.88
                    detailed_analysis:
                      model_statistics:
                        open_questions: 5
                        closed_questions: 2
                        reflections: 3
                      total_counselor_messages: 10
                      prosody_analysis:
                        avg_pitch: 225.3
                        avg_energy: 0.68
                        avg_tempo: 1.15
                      python_api_used: "http://localhost:5000"
        "400":
          description: Validasi gagal
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
              examples:
                not_finished:
                  summary: Session belum selesai
                  value:
                    success: false
                    message: "Session belum selesai"
                already_evaluated:
                  summary: Sudah dievaluasi
                  value:
                    success: false
                    message: "Session ini sudah pernah dievaluasi"
                no_transcript:
                  summary: Tidak ada transkrip
                  value:
                    success: false
                    message: "Session tidak memiliki transkrip"
        "404":
          description: Session tidak ditemukan atau tidak punya akses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
              examples:
                not_found:
                  summary: Session tidak ditemukan
                  value:
                    success: false
                    message: "Session tidak ditemukan"
                no_access:
                  summary: Tidak punya akses
                  value:
                    success: false
                    message: "User tidak punya akses ke session ini"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  details:
                    type: string
                    description: Stack trace (hanya di development mode)
  /reports/stats:
    get:
      tags: [Reports]
      summary: Statistik evaluasi user
      description: >
        Mengambil statistik rata-rata evaluasi untuk user yang sedang login.
        Menghitung rata-rata empathy_score dan question_score dari semua
        session yang sudah dievaluasi.
      responses:
        "200":
          description: Statistik berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Statistik rata-rata berhasil diambil"
                  data:
                    type: object
                    properties:
                      total_evaluations:
                        type: integer
                        description: Total session yang sudah dievaluasi
                        example: 10
                      avg_empathy_score:
                        type: number
                        description: Rata-rata empathy score
                        example: 82.5
                      avg_question_score:
                        type: number
                        description: Rata-rata question score
                        example: 76.3
              examples:
                with_data:
                  summary: User sudah punya evaluasi
                  value:
                    success: true
                    message: "Statistik rata-rata berhasil diambil"
                    data:
                      total_evaluations: 10
                      avg_empathy_score: 82.5
                      avg_question_score: 76.3
                no_data:
                  summary: Belum ada evaluasi
                  value:
                    success: true
                    message: "Belum ada sesi yang dievaluasi"
                    data:
                      total_evaluations: 0
                      avg_empathy_score: 0
                      avg_question_score: 0
        "404":
          description: Session tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sesi tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Gagal mengambil data statistik: error message"
...
  /reports/user/me/statistics:
    get:
      tags: [Reports]
      summary: Statistik lengkap user
      description: >
        Mengambil statistik lengkap untuk user yang sedang login,
        termasuk total sessions, completed/ongoing sessions,
        total hours, dan average scores dari semua evaluasi.
      responses:
        "200":
          description: Statistik user berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSessions:
                    type: integer
                    description: Total semua session user
                    example: 15
                  completedSessions:
                    type: integer
                    description: Total session yang selesai
                    example: 12
                  ongoingSessions:
                    type: integer
                    description: Total session yang masih berjalan
                    example: 3
                  evaluatedSessions:
                    type: integer
                    description: Total session yang sudah dievaluasi
                    example: 10
                  totalHours:
                    type: number
                    description: Total jam konseling
                    example: 8.5
                  averageScores:
                    type: object
                    properties:
                      empathy:
                        type: integer
                        description: Rata-rata empathy score
                        example: 83
                      question:
                        type: integer
                        description: Rata-rata question score
                        example: 77
                      ethics:
                        type: integer
                        description: Rata-rata ethics score
                        example: 0
                      overall:
                        type: integer
                        description: Overall average score
                        example: 53
              examples:
                success:
                  summary: Statistik user lengkap
                  value:
                    totalSessions: 15
                    completedSessions: 12
                    ongoingSessions: 3
                    evaluatedSessions: 10
                    totalHours: 8.5
                    averageScores:
                      empathy: 83
                      question: 77
                      ethics: 0
                      overall: 53
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string